+++ Content.Shared/Audio/AmbientMusicPrototype.cs
@@ -1,4 +1,5 @@
 using Content.Shared.Random;
+using Content.Shared.Random.Rules;
 using Robust.Shared.Audio;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype;
